move plotting functions
